+++ CRM/DataprocessorOutputExport/CSV.php
@@ -165,6 +165,7 @@ class CRM_DataprocessorOutputExport_CSV implements ExportOutputInterface, Direct
 
     $path = CRM_Core_Config::singleton()->templateCompileDir . 'dataprocessor_export_csv/'. $filename;
     if ($sortFieldName) {
+      $dataProcessorClass->getDataFlow()->resetSort();
       $dataProcessorClass->getDataFlow()->addSort($sortFieldName, $sortDirection);
     }
 
@@ -297,6 +298,7 @@ class CRM_DataprocessorOutputExport_CSV implements ExportOutputInterface, Direct
     $dataProcessorClass = \CRM_Dataprocessor_BAO_DataProcessor::dataProcessorToClass($dataProcessor);
     CRM_Dataprocessor_Form_Output_AbstractUIOutputForm::applyFilters($dataProcessorClass, $params);
     if ($sortFieldName) {
+      $dataProcessorClass->getDataFlow()->resetSort();
       $dataProcessorClass->getDataFlow()->addSort($sortFieldName, $sortDirection);
     }
     $dataProcessorClass->getDataFlow()->setOffset($offset);