+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Fixed #24
 * Improved export/import functionality.
 * Added documentation generator to the API output.
+* Added default sort configuration for a data processor (#26).
 
 # Version 1.3.0